fixed memory leaks, need to supress readline leaks

+++ srcs/parsing.c
@@ -3,42 +3,45 @@
 /*                                                        :::      ::::::::   */
 /*   parsing.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aball <aball@student.42.fr>                +#+  +:+       +#+        */
+/*   By: ballzball <ballzball@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 18:22:16 by aball             #+#    #+#             */
-/*   Updated: 2022/11/11 22:41:53 by aball            ###   ########.fr       */
+/*   Updated: 2022/11/12 02:06:05 by ballzball        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
 
-void	check_dir(t_cmd *args)
+int	check_dir(t_cmd *args)
 {
 	args->folder = opendir("/bin");
 	args->dir = readdir(args->folder);
 	while (args->dir)
 	{
-		if (args->dir->d_namlen == ft_strlen(args->cmd[0]) && !ft_strncmp(args->dir->d_name, args->cmd[0], args->dir->d_namlen))
+		if (ft_strlen(args->dir->d_name) == ft_strlen(args->cmd[0]) && !ft_strncmp(args->dir->d_name, args->cmd[0], ft_strlen(args->dir->d_name)))
 		{
 			args->path = ft_strjoin("/bin/", args->cmd[0]);
-			return ;
+			closedir(args->folder);
+			return (1);
 		}
 		args->dir = readdir(args->folder);
 	}
+	closedir(args->folder);
 	args->folder = opendir("/usr/bin");
-	if (!args->folder)
-		printf("fucked\n");
 	args->dir = readdir(args->folder);
 	while (args->dir)
 	{
-		if (args->dir->d_namlen == ft_strlen(args->cmd[0])  && !ft_strncmp(args->dir->d_name, args->cmd[0], args->dir->d_namlen))
+		if (ft_strlen(args->dir->d_name) == ft_strlen(args->cmd[0])  && !ft_strncmp(args->dir->d_name, args->cmd[0], ft_strlen(args->dir->d_name)))
 		{
 			args->path = ft_strjoin("/usr/bin/", args->cmd[0]);
-			return ;
+			closedir(args->folder);
+			return (1);
 		}
 		args->dir = readdir(args->folder);
 	}
 	printf("%s: command not found\n", args->cmd[0]);
+	closedir(args->folder);
+	return (0);
 }
 
 int	parsing(t_cmd *args)
@@ -48,14 +51,13 @@ int	parsing(t_cmd *args)
 	args->line = readline("\x1b[30m\x1b[46mminishell$\x1b[m ");
 	if (!args->line)
 	{
-		free(args->line);
-		exit (1);
+		return (0);
 	}
-	args->expand = (int *)ft_calloc(count_dollar(args->line), sizeof(int));
 	if (args->line && *args->line)
 		add_history(args->line);
 	else
 		return (1);
+	args->expand = (int *)ft_calloc(count_dollar(args->line), sizeof(int));
 	args->cmd = quote_validator(args, 0, 0);
 	if (!args->cmd)
 	{
@@ -67,7 +69,6 @@ int	parsing(t_cmd *args)
 	if (ft_strlen(args->cmd[0]) == 4 && !ft_strncmp(args->cmd[0], "exit", 4))
 	{
 		printf("%s\n", args->cmd[0]);
-		freedom(args->cmd);
 		return (0);
 	}
 	else if (ft_strlen(args->cmd[0]) == 4 && !ft_strncmp(args->cmd[0], "echo", 4))
@@ -82,21 +83,23 @@ int	parsing(t_cmd *args)
 		my_env(args);
 	else if (ft_strlen(args->cmd[0]) == 5 && !ft_strncmp(args->cmd[0], "unset", 5))
 		my_unset(args);
-	else
+	else if (check_dir(args))
 	{
 		pid = fork();
 		if (pid == 0)
 		{
-			check_dir(args);
 			execve(args->path, args->cmd, NULL);
 			exit (1);
 		}
 		else
 		{
+			free(args->path);
 			wait(&pid);
 			kill(pid, SIGQUIT);
 		}
 	}
 	freedom(args->cmd);
+	free(args->line);
+	free(args->expand);
 	return (1);
 }
